Enable downstream method for last partition mapping (#17539)
Users are hitting a `NotImplementedError` when attempting to launch an
asset backfill targeting partitioned assets with a
`LastPartitionMapping` between them.

In this case, when calling `get_downstream_partitions_for_partitions`,
if the upstream subset contains the upstream partitions def's last
partition key, then all downstream partitions are dependencies. This PR
implements this.

One situation this is useful is if you have an upstream time partitioned
asset and a downstream static partitioned asset that incorporates the
last day's data.