feat(dbt): add `get_partition_mapping` to `DagsterDbtTranslator` (#17488)
## Summary & Motivation
Resolves https://github.com/dagster-io/dagster/issues/13053.

Allow `PartitionMapping` to be defined for dbt root dependencies (e.g.
dbt sources, dbt models with no dependencies). In a follow-up PR, we
will add a way to define self-dependencies on dbt models, so that
`PartitionMapping`'s can be defined on a model's previous partitions
rather than the model's inputs.

## How I Tested These Changes
add a test