feat(dbt): allow self-dependencies to be specified in `+meta` (#17490)
## Summary & Motivation
Now that users can specify `PartitionMapping` on their dbt assets, they
can also specify self-dependencies.

In the future, we can explore a generic way of adding arbitrary way of
forcing dependencies using `DagsterDbtTranslator`, but current the use
cases for that are limited (other than forcing self-dependencies). So
for simplicity, make this into a boolean toggle.

Note that dbt's generic version of this is
https://docs.getdbt.com/reference/dbt-jinja-functions/ref#forcing-dependencies.

## How I Tested These Changes
add test