[auto-materialize] Don't store empty evaluation data on the cursor (#17408)
## Summary & Motivation

Two changes here:

1. Stop storing the latest AutoMaterializeAssetEvaluation for an asset
if it is empty, as this wastes a lot of space
2. Update the equivalent_to_stored_evaluation method to return True if
the target evaluation is empty. This will avoid us writing out an
evaluation to the table every tick if an asset continually evaluates as
empty (as we wouldn't have any data in the cursor, but nothing would be
changing tick to tick)

Re: "wastes a lot of space", this doesn't help the worst case scenario
for cursor size, where all assets have at least one partition that is
currently being skipped, but it does help a lot in the more common case
(where most assets do not have any skipped partitions). This took a
1.2Mb cursor down to 4Kb in one test.

## How I Tested These Changes