[auto-materialize] Faster time window partition validation (#17447)
## Summary & Motivation

Two simple changes here, aimed at addressing:
https://github.com/dagster-io/dagster/issues/17429

1. Remove the `_get_validated_time_window_for_partition_key `method, as
this was only used in the context of the `has_partition_key` method, and
neither the body of that function nor the caller actually accessed the
end time for the validated time window. Instead, just compute the start
time for the partition window (much faster in simple cases), and run it
through the same validation steps.
2. When we call `.empty_subset().with_partition_keys()`, stick with the
in-memory format of a list of partition key strings, rather than doing
the expensive conversion process immediately.

I also removed the `is_valid_partition_key` method, as this is
essentially just an outdated version of `has_partition_key` (and does
not actually catch all the cases that `has_partition_key` does, e.g. it
does not consider the end_date of a time partitions def)

## How I Tested These Changes