[ui] Update UI to use new asset checks resolvers, remove old ones (#17591)
## Summary & Motivation

- Switch to using `AssetNode.assetChecksOrError` instead of
`AssetNode.assetChecks` and a top level `assetCheckExecutions` instead
of `assetCheck.executions`?

## How I Tested These Changes

Verify that all the views in Dagster still load and display asset
checks.

I simplified the AssetChecksDetailModal quite a bit because there is no
longer a GraphQL resolver that loads a single check. The checks page now
loads the list of checks and THEN renders the open the modal, which
means that the modal does not need to repeat for the same "migration
required", etc. checks that the checks page does. To verify the behavior
is still correct, I mocked the backend resolver to return the migration
required error and verified i see the error and the modal does not open.

In a follow-up we may want to rename the details modal to
"CheckRunHistoryModal", didn't want to clutter this PR.

Co-authored-by: bengotow <bgotow@elementl.com>