[auto-materialize] Allow updated testing framework to run the AssetDaemon internally (#17428)
## Summary & Motivation

We typically have two modes of testing for the logic. The first is
purely testing the logic (`test_asset_daemon_fast.py`), and the second
is a more "authentic" test (`test_asset_daemon.py`).

Typically, in local development, you'd only run the fast tests, which
don't require invoking heavier daemon machinery, as this will catch
almost all errors. However, you'd still want to make sure that a full
end to end test of the daemon machinery happens in CI/CD.

Note: the current suite of tests runs in ~3 seconds in the "fast" mode,
but ~55 seconds in the "daemon" mode.

This also fixes an issue with the previous framework which made it
impossible to test the daemon against scenarios where the asset graph
changed between ticks, which is a key advantage of this new format (as
it's now much easier to mutate the graph within the context of a
scenario). It does this with a hacky function that elevates values into
the module scope so that they can be picked up as code locations.

Previously, the testing framework would have a single asset graph per
scenario (don't be fooled by definition_change_scenarios.py, this didn't
actually function correctly), and this would remain static.

We do some massaging to make sure that `evaluate_tick()` ends up setting
the same values regardless of if you're calling it from a daemon context
or a "fast" context, meaning all the same assertions can happen
regardless of how this was executed.

## How I Tested These Changes