perf improvements for croniter-less cron iterator
Summary:
By taking advantage of the fact that we know that we're already on the correct hour/day/month/etc. boundary, we can skip unneccesary work and just call pendulum.add for the correct increment.

Test Plan:
BK
Traverse a large graph, see similar performance that we were seeing in master before this round of changes (with the correctness benefits)