[auto-materialize] Port most freshness-policy scenarios to new framework (#17689)
## Summary & Motivation

Further progress on porting over tests from the old framework. Leaves a
few tests in the original `freshness_policy_scenarios.py`.

A few updates to the framework needed to be made in order for us to
accurately test this stuff:

- Need to mock `time.time` so that runs will be executed with the proper
timestamps (pendulum.test() doesn't cover all the places where we get
the current time)
- Need to actually add requested runs to the instance after each tick
evaluation, as freshness logic uses this
- Need a way to request a run that will fail partway through

In several places in the old framework, we'd have patterns where we'd
have one scenario assert that no runs were kicked off if a run had just
failed, then another scenario where it'd be the exact same, but with an
extra delay (and then we'd assert that a run DID get kicked off). This
is much more efficient in the new framework, and so many tests can be
combined into one.

## How I Tested These Changes