fix schedule with job partitions inferred from asset with multi-parti… (#17700)
…tions

## Summary & Motivation

When a partitions def is not explicitly supplied to `define_asset_job`,
we infer one from the partitions on the underlying assets. When we use
`build_schedule_from_partitioned_job` to build a schedule on top of that
job, we use those inferred partitions.

When the inferred partitions definition is a multi-partitions
definition, each schedule tick returns a list containing a run request
per partition.

However, apparently, when you construct a schedule using the
`ScheduleDefinition` constructor instead of the decorator, you're not
allowed to return a list of run requests.

We were using the decorator in the case where the partitions definition
was passed into `define_asset_job`, but using the constructor in the
case where we inferred the partitions definition from the underlying
assets. This fixes the issue by using the decorator in both cases.

Fixes https://github.com/dagster-io/dagster/issues/17592

## How I Tested These Changes