[auto-materialize] Update `test_asset_daemon.py` (#17749)
## Summary & Motivation

Changes:
- a lot of repetitive code in this file for making instances with
different properties was replaced with a single method for that
- stopped switching between using pytest fixtures and inline functions
to create the instances (just use functions, as there isn't much overlap
between tests)
- update the AssetDaemonScenario to add more daemon-specific checks when
running in "daemon mode". this makes several of the existing tests
obsolete, and also helps up our test coverage because we're running
these checks against several different scenarios instead of just one.
- run the full end-to-end daemon against an arbitrary 10% of all
scenarios. on the fence about this one, I think it's probably fine to
just run the end-to-end test against all of the scenarios (just takes a
few minutes)

## How I Tested These Changes