Fix multi-asset subsetting (again) (#15718)
## Summary & Motivation

This offers a new fix for the problem described here:
https://github.com/dagster-io/dagster/pull/13830

The original PR was eventually reverted, as it seemed to be the source
of some mysterious bugs related to the fact that it was generating new
ops on the fly. This PR takes a different approach, wherein we simply
explicitly make new nothing-type inputs when constructing the op in the
first place. This has a couple nice properties:

1. each AssetsDefinition continues to share a single op definition
2. subsettable multi-assets which do not define any internal
dependencies are unchanged
3. we retain a nice input to asset key mapping

This also has a couple drawbacks:

1. people using AssetsDefinition.from_op() will not get access to this
behavior, as you need to explicitly add these dependencies
2. We will be creating a bunch more "useless" inputs for our dbt ops, as
most outputs will now also need to have a corresponding input
representation.
    - Probably not a huge deal

For a visual representation of what's going on here: 

![image](https://github.com/dagster-io/dagster/assets/22457492/7f4f413b-7640-4e2f-b1b4-d4b9f46521fa)

D1 and D2 are both part of the same multi-asset, but needed to get split
up in order for the overall dependency graph to make sense. Previously,
if P was not included in the set of things to execute, there'd be no
dependency chain that forced D2 to execute after D1. This PR adds an
additional input for this purpose.

## How I Tested These Changes

New unit test, plus manually verifying that the correct op graph is now
created