[ui] Add option to “Launch new run” to the job actions menu (#17878)
## Summary & Motivation

Fixes https://github.com/dagster-io/dagster/issues/17627

<img width="457" alt="image"
src="https://github.com/dagster-io/dagster/assets/1037212/80d33efb-6107-43db-a3ca-1b9623a777aa">


For asset jobs, this opens the Materialize modal for all the assets in
the run if partitions or config are required, and immediately launches a
run otherwise. To do this efficiently I made a few modifications to the
`useMaterializationAction` so that you can pass a pipelineSelector
instead of a set of assetKeys.

For op jobs, this takes you to the launchpad where you can add / review
config and click launch run. I think there may be a way to skip the
launchpad if no config is required, but I don't think we have any other
UIs that do a "single-click launch" of fresh op jobs and I'm not sure
this should be the first.

## How I Tested These Changes
Tested this manually and verified that the storybooks still look good.

---------

Co-authored-by: bengotow <bgotow@elementl.com>