[1/n subset refactor] Split TimeWindowPartitionsSubset (#17684)
`TimeWindowPartitionsSubset` has lots of performance improvement code
that relies on representing partitions as keys, but relies on converting
partition keys to time windows in order to serialize. In preparation for
serializing the time window version, this PR splits the implementation:
- `BaseTimePartitionsSubset`: a time partitions subset superclass that
defines abstract methods, and contains helper functions
- `TimePartitionKeyPartitionsSubset`: the "partition key" representation
- `TimeWindowPartitionsSubset`: the "time window" representation

By default, when creating partitions subsets via
`partitions_def.empty_subset().with_partition_keys(...)`, a
`TimePartitionKeyPartitionsSubset` will be created. This class contains
a method to convert to a `TimeWindowPartitionsSubset` when it needs to
be serialized.

In future PRs, time partitions subsets can be directly deserialized as
`TimeWindowPartitionsSubsets`.