Adjust how we describe empty gaps in the AMP timeline
Summary:
We are moving to a world where evaluation IDs don't auto-increment by 1 for each tick, and it's not clear that showing the exact number of evaluations is that useful (esp now that we have a timeline view with per-tick information). So instead of showing "2k evaluations", show "From time A to time B".