[2/n subset refactor] Make TimeWindowPartitionsDefinition serializable (#17660)
This PR supports serializing `TimeWindowPartitionsDefinition` objects.

`whitelist_for_serdes` enforces a constraint where the arguments to
`__new__` must match the ordering of fields in the named tuple. This
constraint is required to pack and unpack objects properly during
pickling.

Unfortunately this ordering doesn't match in
`TimeWindowPartitionsDefinition`, and we can't reorder the fields
without breaking users :( This PR loosens the ordering restriction via a
new `is_pickleable` arg, and raises an error when attempting to pickle a
`TimeWindowPartitionsDefinition`.