test dagster-dbt select with test tags (#17988)
For https://github.com/dagster-io/dagster/issues/17372 - add coverage
for the cases that work currently (deselecting tests, and selecting
specific tests).

What doesn't work is selecting a test without a model. The error is bad-
it falls back to everything. If we decide we'll keep this behavior, I'll
fix up the error case