[3/n subset refactor] Add whitelist_for_serdes to TimeWindowPartitionsSubset (#17702)
This PR converts `TimeWindowPartitionsSubset` to a named tuple decorated
with `@whitelist_for_serdes`. This is mostly field renames (i.e.
`self._included_time_windows -> self.included_time_windows`).

There is a logic change to add a `before_pack` hook in
`whitelist_for_serdes`. This enables mutating the named tuple before it
is serialized, which is used in this PR to force calculating the #
partitions in the `TimeWindowPartitionsSubset`. This has been added
because there is perf logic to delay calculating the # partitions in
`TimeWindowPartitionsSubset` until necessary.

---------

Co-authored-by: Sandy Ryza <sandy@elementl.com>