Only fetch run ids into memory when dequeuing
Previously, we fetched every queued run into memory. Then we sorted
them by priority. Then we applied concurrency limits.

This is fine in the general case, but memory inefficient when there are
large backfills. Imagine a 25k run backfill - all 25k runs need to be
brought into memory at once.

There are a few different ways we can avoid this overfetching. Each has
its own tradeoffs.

- Push all of the concurrency limits down to the database layer. We'd
  need to precompute which tags are immediately disqualified because
  they're already at their concurrency limit and then do some sort of
  "negative" runs filter to match only runs that _don't_ have those
  tags.
- Stream our runs back from the database in order. There are a couple of
  strategies to do this - db server-side cursors, temporary tables,
  writing our own cursoring while holding a db transaction over - but
  all require some amount of care to make sure we have a consistent
  snapshot of the database while we're cursoring.

This change is basically the second one with our in-memory list of
sorted run ids acting as the consistent database snapshot. This
implementation fetches runs one at a time meaning the number of queries
each tick will run is:

_n_+1 where _n_ is the number of runs we need to check until the `batch`
is full. In cases where we have no tag concurrency limits, _n_ will
match the number of runs that get dequeued.

A potential optimization we could make is to paginate through the run
ids list and fetch multiple runs at a time if we want to limit the
number of db queries.