add prettier to makefile (#18045)
Add prettier for formatting yaml. More context:
https://dagsterlabs.slack.com/archives/C03A0D72A6T/p1700081251053029

`npm --prefix ... exec` means we'll use the same version of prettier
(3.0.3) as defined in the dagster-ui project. Eventually we might be
able to put the various prettier steps in to one command, but I think
this is easier to adopt for now. Related, I considered `.prettierignore`
at the root instead of passing files with git ls-files. It was 1, way
slower with the same selection (30s instead of 2). I didn't dig in to
why. And 2, We have different prettier invocations currently. An ignore
file probably makes more sense once we consolidate them