[auto-materialize] Fix issue in which nonexistent parent partitions would be permanently considered nonexistent (#18036)
## Summary & Motivation

Resolves: https://github.com/dagster-io/dagster/issues/18024

Before this PR, once an asset partition was considered to have "required
but non-existent partitions", we would always use the
previously-calculated value for this property (as if there is a large
number of candidate partitions, it's fairly slow to do this partition
mapping process on every tick). However, nonexistent partitions can pop
into existence later on, and so we do need to recalculate.

If an upstream asset partition is required, then it must be the case
that it is materialized in order for the downstream to be kicked off.
Therefore, it's sufficient to not update the state of this value unless
a parent is updated. However, the better way of doing this in the future
would be to add a new context method to detect if a partitions def has
any new partition keys since the previous tick and recalculate this
value whenever a parent gets a new partition key.

## How I Tested These Changes

Added test failed before PR, now passes