Solving issue #17409 with InputContext and Multipartitions (#18109)
## Summary & Motivation
A bug was detected when implementing MultiPartitionsDefinition with one
time-partitioned dimension. The input context was not able to access the
start and end time of the partitioned asset. The issue was reported and
explained here:

https://github.com/dagster-io/dagster/issues/17409

## How I Tested These Changes
The proposed change is easy and it was tested with the same
multipartition implementation that was giving the error before. I can
access the start and end time of the partitioned asset from the input
context, the same as from the output context.

---------

Co-authored-by: Claire Lin <claire@dagsterlabs.com>