Test for invalid priority tags
Sqlite and MySQL both handle invalid priority tags out of the box.

Postgres fails to convert the strings to integers.

This adds a regex replacement to Postgres that will replace any string
that doesn't represent an integer with our default value of 0.

An alternative would be to introduce a custom `try_cast` to Postgres,
but we want this to work even if users don't run our migrations.