[dagster-shell] Add option to exclude shell command from log (#18139)
## Summary & Motivation

This adds the `log_shell_command` option to the `execute` and
`execute_script_file` methods. When set to false, the command will not
be logged to the console, which is useful for commands that contain
sensitive information.

Fixes #14854

## How I Tested These Changes

Added a unit test that fails without the fix and passes with it.