[5/n subset refactor] [serdes] Enable serializing mappings with non-scalar keys (#18057)
This PR enables serializing mappings with non-scalar keys, intended to
be used for serializing mappings keyed by asset key. Previously this was
not possible because `seven.json.dumps` can only serialize dictionaries
keyed by a primitive or a string. Without this change, we would need a
custom `FieldSerializer` when serializing any dict keyed by a
non-scalar.