Only show changed and missing option in node lineage and global asset graph (#18198)
## Summary & Motivation

We used to rely on a field called `liveDataForStale` to determine
whether or not to show this option. Since we stopped fetching all of the
live data upfront we removed that attribute but as a consequence we also
removed the gating of the changed and missing feature. This PR adds back
that gating by introducing a new prop: `showChangedAndMissingOption`

## How I Tested These Changes

locally