[7/n subset refactor] Use new asset backfill data serialization format (#17929)
At long last, we have arrived to the final part in this stack. This PR
migrates `PartitionBackfill` logic to use the new serialization format
of `AssetBackfillData`.

When new backfills are created from now on, the UI will be able to
display the asset backfill page even if the partitions defs are
changed/removed. For old backfills, the UI continue to show the
"partitions def has changed" message, and the asset backfill page will
be blank.

Description of changes:
- Adds an additional `asset_backfill_data` field to `PartitionBackfill`
- Asset backfills from now on will use this field with the new
serialization
- Existing backfills will continue to use
`serialized_asset_backfill_data`. We could force the daemon to migrate
these objects mid-backfill, but that value add is pretty low. It also
improves debug-ability by forcing old backfills to use the old
serialization, and new backfills to use the new serialization.

- Serializes the unique ID of each partitions def in a field on
`AssetBackfillData`. Adds a new method in asset backfill execution that
uses the unique ID to check if partitions defs have changed, in which
case we should stop execution.
- This previously existed in the old serialization version of
`AssetGraphSubset`, but was unfortunately duplicated across each subset
type (materialized, in-progress, failed)

- Adds tests cases to cover this new surface area