[gql] Expose partition mapping on AssetDependency (#18119)
Adds backend support to resolve
https://github.com/dagster-io/dagster/issues/15720

This PR exposes the partition mapping on `GrapheneAssetDependency`,
which represents the "edge" in between an asset and its upstream. The
partition mapping is only returned if it is defined on the asset (and
not inferred). The reasoning here is to avoid introducing complexity in
the UI when the user hasn't provided an explicit definition.

It adds descriptions for partition mappings in case if we'd like to
display that.