add longtext support for event blob / materialization blob in mysql (#18120)
## Summary & Motivation
For arbitrary large metadata blobs in events, the 64kb limit in mysql
text cols is overly restrictive. Converting these to longtext will
provide similar behavior to postgres TEXT columns.

https://github.com/dagster-io/dagster/issues/17982

[GH Discussion](https://github.com/dagster-io/dagster/discussions/18313)

We don't have optional schema changes, so was thinking that I would not
add a schema migration because this would be a pretty expensive,
blocking operation. We can add some guidance to users running into this
to manually migrate the column, but new DBs will be spun up with the
appropriate column types on MySQL.

## How I Tested These Changes
BK