Recommit "add flag to sensor context for first tick after start (#17146)"
For sensors that have been turned off for some time, it's useful to have
some indication so that user code can decide to limit the range of
values to examine (e.g. only look at the last 5 minutes worth of data).

This PR adds the last start time to the stored sensor data, and then
adds a flag to the sensor context by comparing the last tick timestamp
to the last start timestamp.

[INTERNAL_BRANCH=prha/fix_sensor_grpc]

Test Plan:
BK, TBD