[partitions] Create an `AssetSubset` class (#18138)
## Summary & Motivation

Generally, it's difficult to write functions that want to indicate
either:
- An unpartitioned asset should / should not be included in some result
- Some subset of partitions of a partitioned asset should / should not
be included in some result

This is because we use a PartitionsSubset class for indicating subsets
of partitioned assets, but have no equivalent for partitioned assets.

The previous best solution was to make use of the `AssetKeyPartitionKey`
class, which can handle both partitioned and unpartitioned assets.

However, a disadvantage of this class is that it requires us to give up
on any efficiency gains that representing things as raw PartitionSubsets
might gain us -- we always have to call `get_partition_keys()`, which
can be an expensive operation (particularly for the newly-added
`AllPartitionsSubset`, but `TimeWindowPartitionsSubset` is not cheap
either).

This PR implements an `AssetSubset` class which defines operations which
work with both unpartitioned and unpartitioned assets, allowing
`get_partition_keys()` to be called lazily.

~It also updates some internals of the AutoMaterialize logic to make use
of this, as a proof of concept, but I was imagining using this class to
replace most instances of `AbstractSet[AssetKeyPartitionKey]`.~

Looking even farther forward, I'm imagining worlds in which we create
more efficient operators over the underlying PartitionSubset classes.
For example, you could imagine a fancy `__and__` operator on the
`TimeWindowPartitionsSubset` which just intersects the time windows,
rather than needing to find the full set of keys in each subset. In the
current world, it'd be hard to take advantage of this sort of efficiency
gain, as in many call sites we're instantly turning our fancy
`TimeWindowPartitionsSubset` into an `AbstractSet[AssetKeyPartitionKey]`
(one example is the AutoMaterialize code stuff, but we also do this in
the partition mapping world as well).

~This can't be a NamedTuple because we want to override the `__iter__`
method, which breaks all sorts of things on NamedTuples. This also means
that we can't `whitelist_for_serdes` this directly, which is a bit of a
shame, and so an alternate solution will need to be created for
representing these in the serdes layer.~

Re: the above, I realized that if we're not calling this class
`AssetPartitions`, then it's not all that surprising that `__iter__`
doesn't yield a stream of AssetKeyPartitionKeys. It is still definitely
a bummer that `__len__` can't be implemented either, but that's a small
price to pay for the ability to `@whitelist_for_serdes` this (IMO)

## How I Tested These Changes