[auto-materialize] Frontload parent updated queries (#18189)
## Summary & Motivation

This helps avoid some (but not all) of the risk of missing a parent
update. Imagine you have an asset that is evaluated near the beginning
of a tick, and no updated parents are found. During the course of the
tick, one of the parents updates. At the end of the tick, the cursor
will be moved past that parent update, and it will be lost.

This helps minimize this risk by minimizing the amount of time between
the parent update calculation and the cursor advancement.

To totally fix this problem, we'll want to store a per-asset cursor so
that things can be updated independently, so this is more of a temporary
solution.

## How I Tested These Changes