feat(dbt): create manifest in static target path at run time (#18328)
## Summary & Motivation
With https://github.com/dagster-io/dagster/issues/16951, we now have the
tools to re-parse the dbt manifest in a static directory.

Invocations of dbt commands in asset or job executions should continue
to use unique dbt target paths to retain references to dbt artifacts.
But for local development, the dbt target path should be static to:

- Reduce confusion about the necessity of the unique target path, as its
presence is exposed early on in local development. This should only be
known at materialization time.
- Reduce the memory footprint of local development. Since the Dagster
project is reloaded every so often, the previous behavior was that the
dbt project's `target/` folder could be clogged up with all the
re-parsed manifests.

## How I Tested These Changes
pytest, local