[auto-materialize] enforce monotonic storage id (#18375)
## Summary & Motivation

We changed how we get the latest_storage_id for the cursor to only look
at the maximum materialization storage id across the assets. In a
previous commit, it was looking for the maximum storage id across all
events. If you serialized a cursor with a high storage id, then changed
to the new system, you'd cause an error because we always expect the
storage id to increase

## How I Tested These Changes