Changing the first example to a functional api. (#18141)
## Summary & Motivation

The current example for partitioned assets is fictitious and does not
run. This is confusing.
This change swaps out the example for one that can be run and provides a
daily partitioned asset against a public NASA API that accepts
`api_key=DEMO_KEY`

## How I Tested These Changes

I implemented the code in a local instance of DAGSTER using the most
recent version (1.5.9). And materialized the asset, and ran the
backfill.
I updated the docs document and proofed it locally in .mdx format.

<img width="308" alt="image"
src="https://github.com/dagster-io/dagster/assets/8226282/ef308293-f83c-44aa-baa7-6c8441bb583c">

---------

Co-authored-by: Erin Cochran <erin.k.cochran@gmail.com>