Fix stuck backfill when definition is changed (#18378)
Previously, when the `DAGSTER_BACKFILL_RETRY_DEFINITION_CHANGED_ERROR`
env var is set, we never mark the backfill as failed regardless of
whether any code location is actually unloadable. This caused certain
backfills to be stuck forever in a canceling state when the partitions
defs were changed, when they should have been marked as failed.

This PR fixes the issue.