Implement a sync version of the schedule and sensor grpc call (#18261)
## Summary & Motivation
We've never needed this to be an async call and we seem to be hitting
upstream grpc issues with streaming API calls where they sporadically
return early without raising an error. Make a sync version of the call
and use that. For back-compat, fall back to the async version if an
UnimplementedError is returned.

This adds a marginal bit of latency in the case where you update your
webserver/daemon/dagster cloud agent before you upgrade your code, but
shouldn't increase the throughput on the boxes since the first call
returns more or less immediately with an UnimplementedError without
doing any actual work. If that's a concern we could try to add some
caching of the fact that a particular server doesn't have the method
implemented, but my sense is that that is more trouble than it is worth.

## How I Tested These Changes
BK, test back-compat case manually by running a sensor locally on a new
version of dagster-webserver but an old version of `dagster api grpc'